Display architecture update, part 1  Use floating-point coordinates  Avoid using G_plot_* for display  Drivers don't need argc/argv  Miscellaneous clean-up
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@32650 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ display/d.barscale/draw_scale.c
@@ -65,11 +65,11 @@ int draw_scale(int toptext)
     double line_len;
     int incr;
     int x_pos, y_pos;
-    int t, b, l, r;
+    double t, b, l, r;
     int pt, pb, pl, pr;
     int i;
     int size;
-    int xarr[5], yarr[5];
+    double xarr[5], yarr[5];
     double seg_len;
     const struct scale *scales = all_scales[use_feet];
