+++ lib/cairodriver/Draw_point.c
@@ -3,12 +3,13 @@
 #define POINTSIZE 1.0
 #define HALFPOINTSIZE (0.5*POINTSIZE)
 
-void Cairo_draw_point(int x, int y)
+void Cairo_draw_point(double x, double y)
 {
-    G_debug(3, "Cairo_draw_point: %d %d", x, y);
+    G_debug(3, "Cairo_draw_point: %f %f", x, y);
 
-    cairo_rectangle(cairo, (double)x - HALFPOINTSIZE,
-		    (double)y - HALFPOINTSIZE, POINTSIZE, POINTSIZE);
+    cairo_rectangle(cairo,
+		    x - HALFPOINTSIZE, y - HALFPOINTSIZE,
+		    POINTSIZE, POINTSIZE);
     cairo_fill(cairo);
     modified = 1;
 }
