+++ lib/cairodriver/Poly.c
@@ -1,6 +1,6 @@
 #include "cairodriver.h"
 
-void do_polygon(const int *xarray, const int *yarray, int count)
+void do_polygon(const double *xarray, const double *yarray, int count)
 {
     int i;
 
@@ -9,14 +9,14 @@ void do_polygon(const int *xarray, const int *yarray, int count)
 	cairo_line_to(cairo, xarray[i], yarray[i]);
 }
 
-void Cairo_Polygon(const int *xarray, const int *yarray, int count)
+void Cairo_Polygon(const double *xarray, const double *yarray, int count)
 {
     G_debug(3, "Cairo_Polygon (%d points)", count);
     do_polygon(xarray, yarray, count);
     cairo_fill(cairo);
 }
 
-void Cairo_Polyline(const int *xarray, const int *yarray, int count)
+void Cairo_Polyline(const double *xarray, const double *yarray, int count)
 {
     G_debug(3, "Cairo_Polyline (%d points)", count);
     do_polygon(xarray, yarray, count);
