+++ lib/cairodriver/Set_window.c
@@ -1,11 +1,10 @@
 #include "cairodriver.h"
 
-void Cairo_Set_window(int t, int b, int l, int r)
+void Cairo_Set_window(double t, double b, double l, double r)
 {
-    G_debug(1, "Cairo_Set_window: %d %d %d %d", t, b, l, r);
+    G_debug(1, "Cairo_Set_window: %f %f %f %f", t, b, l, r);
 
     cairo_reset_clip(cairo);
-    cairo_rectangle(cairo, (double)l, (double)t, (double)r - l,
-		    (double)b - t);
+    cairo_rectangle(cairo, l, t, r - l, b - t);
     cairo_clip(cairo);
 }
