+++ lib/driver/Draw.c
@@ -9,13 +9,13 @@ void DRV_draw_bitmap(int ncols, int nrows, int threshold,
 	(*driver->draw_bitmap) (ncols, nrows, threshold, buf);
 }
 
-void DRV_draw_line(int x0, int y0, int x1, int y1)
+void DRV_draw_line(double x0, double y0, double x1, double y1)
 {
     if (driver->draw_line)
 	(*driver->draw_line) (x0, y0, x1, y1);
 }
 
-void DRV_draw_point(int x, int y)
+void DRV_draw_point(double x, double y)
 {
     if (driver->draw_point)
 	(*driver->draw_point) (x, y);
