+++ lib/driver/Move.c
@@ -1,13 +1,13 @@
 #include "driver.h"
 #include "driverlib.h"
 
-void COM_Move_abs(int x, int y)
+void COM_Move_abs(double x, double y)
 {
     cur_x = x;
     cur_y = y;
 }
 
-void COM_Move_rel(int x, int y)
+void COM_Move_rel(double x, double y)
 {
     cur_x += x;
     cur_y += y;
