+++ lib/psdriver/Draw_line.c
@@ -1,7 +1,7 @@
 
 #include "psdriver.h"
 
-void PS_draw_line(int x1, int y1, int x2, int y2)
+void PS_draw_line(double x1, double y1, double x2, double y2)
 {
-    output("%d %d %d %d LINE\n", x1, y1, x2, y2);
+    output("%f %f %f %f LINE\n", x1, y1, x2, y2);
 }
